# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenant": {
      "get": {
        "operationId": "TenantsController_getTenant",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dados da empresa",
        "tags": [
          "tenant"
        ]
      },
      "patch": {
        "operationId": "TenantsController_updateTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "legalName": {
                    "type": "string",
                    "maxLength": 200,
                    "nullable": true
                  },
                  "document": {
                    "type": "string",
                    "maxLength": 20,
                    "nullable": true
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 30,
                    "nullable": true
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "nullable": true
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "maxLength": 200
                      },
                      "number": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "complement": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "district": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "city": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "state": {
                        "type": "string",
                        "maxLength": 2
                      },
                      "zip": {
                        "type": "string",
                        "maxLength": 10
                      }
                    },
                    "additionalProperties": false
                  },
                  "primaryColor": {
                    "type": "string",
                    "pattern": "^#[0-9a-fA-F]{6}$",
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualiza dados da empresa",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/onboarding/complete": {
      "post": {
        "operationId": "TenantsController_completeOnboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Conclui o onboarding da assistência",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/usage": {
      "get": {
        "operationId": "TenantsController_usage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Uso das quotas do plano (filiais, técnicos e caixas por filial)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/audit": {
      "get": {
        "operationId": "TenantsController_audit",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "entity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 60
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trilha de auditoria da empresa",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/support-access": {
      "get": {
        "operationId": "TenantsController_listSupport",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Acessos de suporte concedidos",
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "TenantsController_grantSupport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "platformUserEmail": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "hours": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 72
                  }
                },
                "required": [
                  "platformUserEmail",
                  "reason",
                  "hours"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Autoriza acesso temporário e auditado do suporte (somente proprietário)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/support-access/{id}": {
      "delete": {
        "operationId": "TenantsController_revokeSupport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoga acesso de suporte",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/branches": {
      "get": {
        "operationId": "TenantsController_listBranches",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Filiais autorizadas",
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "TenantsController_createBranch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 30,
                    "nullable": true
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 64,
                    "default": "America/Sao_Paulo"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "maxLength": 200
                      },
                      "number": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "complement": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "district": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "city": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "state": {
                        "type": "string",
                        "maxLength": 2
                      },
                      "zip": {
                        "type": "string",
                        "maxLength": 10
                      }
                    },
                    "additionalProperties": false
                  },
                  "businessHours": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 40
                    }
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria filial (respeita limite do plano)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/branches/{id}": {
      "get": {
        "operationId": "TenantsController_getBranch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe da filial",
        "tags": [
          "tenant"
        ]
      },
      "patch": {
        "operationId": "TenantsController_updateBranch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 30,
                    "nullable": true
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 64,
                    "default": "America/Sao_Paulo"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "maxLength": 200
                      },
                      "number": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "complement": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "district": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "city": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "state": {
                        "type": "string",
                        "maxLength": 2
                      },
                      "zip": {
                        "type": "string",
                        "maxLength": 10
                      }
                    },
                    "additionalProperties": false
                  },
                  "businessHours": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 40
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualiza/ativa/desativa filial",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members": {
      "get": {
        "operationId": "TenantsController_listMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Membros e convites pendentes",
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "TenantsController_createMember",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "TENANT_OWNER",
                      "TENANT_ADMIN",
                      "MANAGER",
                      "RECEPTIONIST",
                      "TECHNICIAN",
                      "CASHIER",
                      "INVENTORY"
                    ]
                  },
                  "branchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50,
                    "default": []
                  },
                  "technicianBranchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50,
                    "default": []
                  }
                },
                "required": [
                  "email",
                  "role"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convida membro (alias de /members/invite)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members/technicians": {
      "get": {
        "operationId": "TenantsController_technicians",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Técnicos ativos (por filial)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members/invite": {
      "post": {
        "operationId": "TenantsController_invite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "TENANT_OWNER",
                      "TENANT_ADMIN",
                      "MANAGER",
                      "RECEPTIONIST",
                      "TECHNICIAN",
                      "CASHIER",
                      "INVENTORY"
                    ]
                  },
                  "branchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50,
                    "default": []
                  },
                  "technicianBranchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50,
                    "default": []
                  }
                },
                "required": [
                  "email",
                  "role"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convida membro por e-mail (expira em 7 dias)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members/invitations/{id}": {
      "delete": {
        "operationId": "TenantsController_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoga convite",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members/{id}": {
      "patch": {
        "operationId": "TenantsController_updateMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "TENANT_OWNER",
                      "TENANT_ADMIN",
                      "MANAGER",
                      "RECEPTIONIST",
                      "TECHNICIAN",
                      "CASHIER",
                      "INVENTORY"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "DISABLED"
                    ]
                  },
                  "branchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50
                  },
                  "technicianBranchIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 50
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Altera papel, status e filiais do membro",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/members/transfer-ownership": {
      "post": {
        "operationId": "TenantsController_transfer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "membershipId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                },
                "required": [
                  "membershipId",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transfere a propriedade da empresa (confirmação por senha)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "operationId": "TenantsController_getSettings",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Configurações efetivas da empresa/filial",
        "tags": [
          "tenant"
        ]
      },
      "put": {
        "operationId": "TenantsController_putSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entries": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "pattern": "^[a-z0-9_.]{2,80}$"
                        },
                        "value": {},
                        "branchId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        }
                      },
                      "required": [
                        "key"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 30
                  }
                },
                "required": [
                  "entries"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualiza configurações",
        "tags": [
          "tenant"
        ]
      }
    },
...
  },
  "info": {
    "title": "OrdemCerta API",
    "description": "API REST v1 — gestão de assistências técnicas. Erros: {code,message,details,requestId}.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}